+++ Changes.md
@@ -1,3 +1,4 @@
+* #828 Remove the concept of public finalist categories
 * #818 - Fix setting of custom judging and award groups
 
 Release 14.4.0